trying to fix cart

+++ app/src/main/java/com/example/kamuko/DBCursor.java
@@ -56,7 +56,7 @@ public class DBCursor extends CursorWrapper {
     {
         String id = getString(getColumnIndex(DBSchema.cartTable.Cols.ID));
         int count = getInt(getColumnIndex(DBSchema.cartTable.Cols.COUNT));
-        double price = getInt(getColumnIndex(DBSchema.cartTable.Cols.PRICE));
+        double price = getDouble(getColumnIndex(DBSchema.cartTable.Cols.PRICE));
         return new Cart(id, count, price);
     }
 }
